Recurse through full query tree & close channels when done or cancelled
This is effectively a breadth-first search, with the addition of cancellation support.
Crucially, any individual call to 'streamTree' does not have enough context to close any channels.
Instead, we run a loop which can detect when there is nothing left to process.

It would be nicely to make this parallel, to mask the time cost of fetching content over the network,
but that introduces complexity in terms of how to know all subtrees have been fully explored.