Remove types for deeply-nested GraphQL tree query
The new, recursive streaming approach to exploring the repo's tree uses a single node at a time.
As such, it is inefficient to query multiple levels.
This design decision is also motivated by the fact that a single node could have tens, hundreds,
or possibly even thousands of subdirectories; trying to query multiple levels in one go could
prove hideously expensive, to the point of blocking channels or consuming unacceptable levels of
memory on users' machines.